refactor(loon-core): split namespace administration modules
