+++ crates/loon-core/src/options.rs
@@ -1,5 +1,6 @@
 use crate::{
-    basis::VerifiedNamespaceBasis, checkpoint::MetadataTableCache, path::write::PutFileBehavior,
+    checkpoint::MetadataTableCache, namespace::basis::VerifiedNamespaceBasis,
+    path::write::PutFileBehavior,
 };
 use loon_api::wire::control::HeadState;
 use loon_api::CommitId;
